WIP: trau_frame: New API
The old TRAU frame code in src/trau_frame.c (which is for some strange
reason part of libosmo-abis, and not libosmo-trau!) was introduced more
than a decade ago for the needs of bs11_abis AKA OpenBSC aka OsmoNITB.

It is too constrained to implement TRAU frame parsing in a generic way,
including:
* no way to add support for 8k sub-slots
* no way to handle CRC bits or UFI (i.e. no AMR)
* no real API design, it was just ripped out from OsmoNITB and moved into
  a library (even the wrong one)

For those reasons, let's introduce a new API for TRAU frame
encoding/decoding - one that supports all the relevant use cases.

This is marked WIP until the part handling T-bits is properly
implemented.

Change-Id: I5cf42e6c445d9224be18503cebc7584b3beba08c